removed a TODO
This was about a fast drop-in replacement plugin for `filetype.vim`.
That plugin only worked for Neovim, and Neovim has since gotten a
`filetype.lua` which is blazingly fast too, so I guess there's no need
for the plugin anymore, unless perhaps for a narrow range of older
Neovim versions.